Exceptions og billeder

+++ Pages/Sailors/DeleteSailor.cshtml.cs
@@ -13,18 +13,33 @@ namespace SejlKlubsApp.Pages.Sailors
     {
         [BindProperty]
         public Sailor Sailor { get; set; }
-        ISailorService sailorService;
+        private readonly ISailorService sailorService;
+        public string InfoText { get; set; }
         public DeleteSailorModel(ISailorService service)
         {
             sailorService = service;
         }
-        public async Task OnGetAsync(int id)
+        public async Task<IActionResult> OnGetAsync(int id)
         {
+            InfoText = "Du sletter en sejler! Er du sikker?";
             Sailor = await sailorService.GetSailorByIdAsync(id);
+            return Page();
         }
         public async Task<IActionResult> OnPostAsync()
         {
-            await sailorService.DeleteSailorAsync(Sailor);
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+            try
+            {
+                await sailorService.DeleteSailorAsync(Sailor);
+            }
+            catch(Exception e)
+            {
+                InfoText = $"Noget gik galdt! {e.Message}";
+                return Page();
+            }            
             return RedirectToPage("GetAllSailors");
         }
     }
